+++ modules/Students/includes/General_Info.inc.php
@@ -138,20 +138,17 @@ echo PasswordInput(
 		|| ! empty( $_REQUEST['moodle_create_student'] ) ? '' : str_repeat( '*', 8 ) ),
 	'students[PASSWORD]',
 	_( 'Password' ) .
-	( ! empty( $_REQUEST['moodle_create_student'] )
-		|| ! empty( $old_student_in_moodle ) ?
+	( ! empty( $_REQUEST['moodle_create_student'] ) ?
 		'<div class="tooltip"><i>' .
 		_( 'The password must have at least 8 characters, at least 1 digit, at least 1 lower case letter, at least 1 upper case letter, at least 1 non-alphanumeric character' ) .
 		// @since 5.9 Moodle creates user password if left empty.
-		( ! empty( $_REQUEST['moodle_create_student'] ) ?
-			'. ' ._( 'Moodle will create a password and send an email to user if left empty.' ) :
-			'' ) .
+		'. ' ._( 'Moodle will create a password and send an email to user if left empty.' ) .
 		'</i></div>' :
 		''
 	),
 	'maxlength="42" strength' .
-	// @since 5.9 Moodle creates user password if left empty.
-	( ! $required || ! empty( $_REQUEST['moodle_create_student'] ) ? '' : ' required' ),
+	// @since 5.9 Moodle creates user password if left empty + Do not update Moodle user password.
+	( basename( $_SERVER['PHP_SELF'] ) == 'index.php' ? ' required' : '' ),
 	empty( $_REQUEST['moodle_create_student'] )
 );